# server/swagger.yaml
swagger: '2.0'
info:
  description: Mahboonkrong Rice Library Api
  version: 1.0.0
  title: Mahboonkrong Rice
schemes:
  - https
  - http
paths:
  /mbkserver/members:
    post:
      tags:
        - Members
      summary: Add a new Members
      description: ''
      operationId: createMembers
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
             $ref: '#/definitions/tbMembers'
      responses:
        '200':
          description: Success
        '405':
          description: Invalid input
    put:
      tags:
        - Members
      summary: Update an existing Members
      description: ''
      operationId: updateMembers
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            $ref: '#/definitions/tbMembers'
      responses:
        '200':
          description: Success
        '404':
          description: Members not found
    get:
      tags:
        - Members
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      produces:
        - application/xml
        - application/json
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: true
          type: array
          items:
            type: string
            enum:
              - available
              - pending
              - sold
            default: available
          collectionFormat: multi
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/tbMembers'
        '400':
          description: Invalid status value
  /mbkserver/members/Show:
    get:
      tags:
        - Members
      summary: Finds Pets by tags
      description: >-
        Muliple tags can be provided with comma separated strings. Use        
        tag1, tag2, tag3 for testing.
      operationId: findPetsByTags
      produces:
        - application/xml
        - application/json
      parameters:
        - name: tags
          in: query
          description: Tags to filter by
          required: true
          type: array
          items:
            type: string
          collectionFormat: multi
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/tbMembers'
        '400':
          description: Invalid tag value
  /mbkserver/members/export:
    get:
      tags:
        - Members
      summary: Find pet by ID
      description: Returns a single pet
      operationId: getPetById
      produces:
        - application/xml
        - application/json
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: true
          type: array
          items:
            type: string
            enum:
              - available
              - pending
              - sold
            default: available
          collectionFormat: multi
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMembers'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  /mbkserver/members/getMember:
    get:
      tags:
        - Members
      summary: Find member by ID
      description: Returns a single member
      operationId: getMember
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: integer
                format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMembers'
        '400':
          description: Invalid ID supplied
        '404':
          description: Member not found
  /mbkserver/members/getMemberPoints:
    get:
      tags:
        - Members
      summary: People find employee points that will expire within a year.
      description: Returns Member & Member Point
      operationId: getMemberPoint
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: integer
                format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMemberPoints'
        '404':
          description: Member not found
  /mbkserver/members/getMemberPointsList:
    get:
      tags:
        - Members
      summary: Retrieve employee scores by member ID.
      description: Returns Member & Member Point By memberId
      operationId: getMemberPointsList
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: integer
                format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMemberPoints'
        '404':
          description: Member not found
  /mbkserver/members/byId/{id}:
    get:
      tags:
        - Members
      summary: Find purchase member by ID
      description: >-
        For valid response try integer IDs with value >= 1 and <= 10.        
        Other values will generated exceptions
      operationId: getMemberById
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of member that needs to be fetched
          required: true
          type: integer
          maximum: 10
          minimum: 1
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMembers'
        '404':
          description: Members not found
  /mbkserver/members/Show/byId/{id}:
    get:
      tags:
        - Members
      summary: Find purchase member by ID
      description: >-
        For valid response try integer IDs with value >= 1 and <= 10.        
        Other values will generated exceptions
      operationId: getMemberShowById
      produces:
        - application/xml
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of member that needs to be fetched
          required: true
          type: integer
          maximum: 10
          minimum: 1
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/tbMembers'
        '404':
          description: Members not found
  /mbmserver/members/byUid:
    post:
      tags:
        - Members
      summary: Updates a Member
      description: ''
      operationId: updateMemberById
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            type: object
            required:
              - uid
            properties:
              uid:
                type: string
                example: uid9xxxxx51564564561
      responses:
        '405':
          description: Invalid input   
  /mbmserver/members/checkRegister:
    post:
      tags:
        - Members
      summary: Updates a Member
      description: ''
      operationId: checkRegister
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            type: object
            required:
              - uid
            properties:
              uid:
                type: string
                example: uid9xxxxx51564564561
      responses:
        '405':
          description: Invalid input   
  /mbkserver/members/{memberId}:
    delete:
      tags:
        - Members
      summary: Delete Member by ID
      description: >-
        For valid response try integer IDs with positive integer value.        
        Negative or non-integer values will generate API errors
      operationId: deleteMemberById
      produces:
        - application/xml
        - application/json
      parameters:
        - name: memberId
          in: path
          description: ID of the member that needs to be deleted
          required: true
          type: integer
          minimum: 1
          format: int64
      responses:
        '200':
          description: Success
        '400':
          description: Invalid ID supplied
        '404':
          description: member not found
...
definitions:
...
  tbMembers:
    type: object
    required:
      - firstName
      - lastName
      - phone
      - email
      - birthDate
      - registerDate
    properties:
      id:
        type: integer
        example: 1
      firstName:
        type: string
        example: doggie
      lastName:
        type: string
        example: boonrang
      phone:
        type: string
        example: 080-8888888
      email:
        type: string
        example: www@gmail.com
      birthDate: 
        type: string
        example: 09/09/2019
      registerDate:
        type: string  
        example: 09/09/2019
      memberPoint:
        type: integer  
        example: 100
      memberPointExpire:
        type: string  
        example: 09/09/2019
    xml:
      name: tbMembers
  tbMemberPoints:
    type: object
    required:
      - tbMemberId
    properties:
      id:
        type: integer
        format: int64
      tbMemberId:
        type: integer
        format: int64
      tbPointCodeHDId:
        type: integer
        format: int64
      campaignType:
        type: integer
        format: int64
      code:
        type: string
        example: PRGC-xx1xx51x
      point:
        type: integer
        example: 1200
      redeemDate: 
        type: string
        example: 09/09/2019
      expireDate: 
        type: string
        example: 09/09/2019
    xml:
      name: tbMemberPoints
...